requirements.txt
asgiref==3.5.2
certifi==2021.10.8
charset-normalizer==2.0.9
cloudinary==1.30.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
greenlet==2.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==6.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==9.3.0
psycopg2==2.9.5
PyJWT==2.6.0
pyotp==2.8.0
pyparsing==3.0.6
python-decouple==3.7
python-emailer==1.5.3
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.2
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.12.0
